feat: abstract types into relevant files to maintain readability
