+++ src/sections/About/About.tsx
@@ -1,10 +1,10 @@
 import { useRef } from 'react';
-import type { MutableRefObject } from 'react';
+// import type { MutableRefObject } from 'react';
 
-import { styled } from '@mui/material/styles';
-import { Avatar, Box, Badge, Typography } from '@mui/material';
+import { Avatar, Box, Typography } from '@mui/material';
 
 import { lightPaperStyle } from '../../pages';
+import { ImageWrapper, StyledBadge } from './AboutMUIStyles';
 
 import { useIntersectionObserver } from '../../utils';
 
@@ -20,45 +20,9 @@ const introductionData = {
    I am having fun in React and RESTing in between.`,
 };
 
-const StyledBadge = styled(Badge)(({ theme }) => ({
-  '& .MuiBadge-badge': {
-    backgroundColor: '#44b700',
-    color: '#44b700',
-    boxShadow: `0 0 0 2px ${theme.palette.background.paper}`,
-    '&::after': {
-      position: 'absolute',
-      top: 0,
-      left: 0,
-      width: '100%',
-      height: '100%',
-      borderRadius: '50%',
-      animation: 'ripple 1.2s infinite ease-in-out',
-      border: '1px solid currentColor',
-      content: '""',
-    },
-  },
-  '@keyframes ripple': {
-    '0%': {
-      transform: 'scale(.8)',
-      opacity: 1,
-    },
-    '100%': {
-      transform: 'scale(2.4)',
-      opacity: 0,
-    },
-  },
-}));
-
-const ImageWrapper = styled(Box)(({ theme }) => ({
-  float: 'left',
-  marginTop: theme.spacing(1),
-  marginRight: theme.spacing(2),
-  marginBottom: theme.spacing(1),
-}));
-
 
 export const About: React.FC = () => {
-  const pictureRef = useRef<HTMLElement | null>(null);
+  const pictureRef = useRef<HTMLElement | undefined>(undefined);
   const isInView = useIntersectionObserver(pictureRef);
 
   return (
