+++ src/sections/About/AboutMUIStyles.ts
@@ -1,0 +1,38 @@
+import { styled } from '@mui/material/styles';
+import { Box, Badge } from '@mui/material';
+
+export const StyledBadge = styled(Badge)(({ theme }) => ({
+    '& .MuiBadge-badge': {
+      backgroundColor: '#44b700',
+      color: '#44b700',
+      boxShadow: `0 0 0 2px ${theme.palette.background.paper}`,
+      '&::after': {
+        position: 'absolute',
+        top: 0,
+        left: 0,
+        width: '100%',
+        height: '100%',
+        borderRadius: '50%',
+        animation: 'ripple 1.2s infinite ease-in-out',
+        border: '1px solid currentColor',
+        content: '""',
+      },
+    },
+    '@keyframes ripple': {
+      '0%': {
+        transform: 'scale(.8)',
+        opacity: 1,
+      },
+      '100%': {
+        transform: 'scale(2.4)',
+        opacity: 0,
+      },
+    },
+  }));
+  
+export const ImageWrapper = styled(Box)(({ theme }) => ({
+    float: 'left',
+    marginTop: theme.spacing(1),
+    marginRight: theme.spacing(2),
+    marginBottom: theme.spacing(1),
+  }));
