+++ src/sections/Contacts/Contact.tsx
@@ -1,18 +1,12 @@
+import React from 'react';
+
 import { SiGithub, SiLinkedin } from 'react-icons/si';
-import { IconType } from 'react-icons';
 
 import { Box } from '@mui/material';
 
 import { IconWithTooltip } from '../../components/index';
-import React from 'react';
+import type { Contacts, ContactKey } from '../../types/contactTypes';
 
-
-interface Contact {
-  icon: IconType,
-  url: string,
-}
-
-type Contacts = Record<string, Contact>;
 
 const contactsData: Contacts = {
   Github: {
@@ -25,11 +19,10 @@ const contactsData: Contacts = {
   },
 };
 
-type ContactKey = keyof typeof contactsData;
-
 type Props = {
   contact: ContactKey,
 }
+
 
 export const Contact: React.FC<Props> = ({ contact }) => {
   const { icon, url } = contactsData[contact];
