+++ src/types/contactTypes.ts
@@ -1,0 +1,11 @@
+import { IconType } from 'react-icons';
+
+interface Contact {
+    icon: IconType,
+    url: string,
+  }
+  
+export type Contacts = Record<string, Contact>;
+
+export type ContactKey = 'Github' | 'LinkedIn';
+
