+++ src/types/projectTypes.ts
@@ -2,10 +2,10 @@ export interface Project {
         title: string,
         subheader: string,
         githubLink: string,
-        liveLink: string,
-        video: string,
         thumbnails: string,
         details: string[],
+        liveLink?: string,
+        video?: string,
 }
 
 export type Projects = Record<number, Project>;
